requirements.txt
google-generativeai
Pillow
streamlit
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
langchain-community
chromadb